changed float variables to double variables

+++ src/main/java/com/midaswebserver/midasweb/forms/BackTraderOptimizeForm.java
@@ -35,13 +35,13 @@ public class BackTraderOptimizeForm {
     @Min(0)
     @Max(1)
     @NotBlank(message = "Please enter stake in decimal form (3% -> .03)")
-    public float stake;
+    public double stake;
     @NotNull
     @NotBlank
     @Min(0)
     @Max(1)
     @NotBlank(message = "Please enter commission in decimal form (3% -> .03)")
-    public float commission;
+    public double commission;
     @NotNull
     @NotBlank
     @Max(100)
@@ -89,19 +89,19 @@ public class BackTraderOptimizeForm {
         this.algorithm = algorithm;
     }
 
-    public float getStake() {
+    public double getStake() {
         return stake;
     }
 
-    public void setStake(float stake) {
+    public void setStake(double stake) {
         this.stake = stake;
     }
 
-    public float getCommission() {
+    public double getCommission() {
         return commission;
     }
 
-    public void setCommission(float commission) {
+    public void setCommission(double commission) {
         this.commission = commission;
     }
